refactor: split ClaimPage.tsx into focused sub-components
Extract step-specific UI into dedicated components:
- ClaimStepIndicator: step progress bar
- ClaimEmailStep: email + magic link form
- ClaimCheckEmailStep: check-inbox screen with resend
- ClaimWebsiteStep: website URL input
- ClaimVerifyStep: verification link display
- ClaimManualReviewStep: manual review request form
- ClaimManualReviewSubmittedStep: confirmation screen
- ClaimReviewStep: review links, photo, location
- ClaimDoneStep: success screen

Shared extracted:
- ClaimPageTypes: ClaimStep union, ReviewLink interface
- ClaimPageUtils: AVATAR_PLATFORMS, platformName helper

ClaimPage: 765 → 432 lines
All location feature (city/country) preserved.

Fixes UNS-21
